Make flashcards to separate features folder
