+++ src/features/flashcards/components/Flashcards.tsx
@@ -3,9 +3,9 @@ import React, { useEffect, useState, useCallback, useMemo } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { useParams } from 'react-router-dom';
 import { RootState } from 'typesafe-actions';
-import Spinner from '../../../../layout/components/Spinner';
-import useNavBackRoute from '../../../../layout/hooks/useNavBackRoute';
-import { fetchArticleAsync } from '../../actions';
+import Spinner from '../../../layout/components/Spinner';
+import useNavBackRoute from '../../../layout/hooks/useNavBackRoute';
+import { fetchArticleAsync } from '../../content/actions';
 import FlashcardBody from './FlashcardBody';
 import FlashcardHeader from './FlashcardHeader';
 import FlashcardsController from './FlashcardsController';
