+++ src/store/root-reducer.ts
@@ -1,7 +1,7 @@
 import { combineReducers } from 'redux';
 import auth from '../features/auth/reducer';
 import content from '../features/content/reducer';
-import flashcards from '../features/content/components/flashcards/reducer';
+import flashcards from '../features/flashcards/reducer';
 import search from '../features/search/reducer';
 import settings from '../features/settings/reducer';
 import layout from '../layout/reducer';
